Refactored Base64 decoder. Split into different methods.
