+++ lib/tools/json_beautify/json_beautify.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:glad_tools/components/ui/bordered_all.dart';
-import 'package:glad_tools/models/json_beautify/json_parser_isolate.dart';
-import 'package:glad_tools/models/tool_object.dart';
+import 'package:glad_tools/tools/json_beautify/json_parser_isolate.dart';
+import 'package:glad_tools/tools/tool_object.dart';
 
 class JsonBeautify extends ToolObject {
   static dynamic rootObject;
