+++ lib/tools/url_inspector/headers_view.dart
