+++ lib/views/main_view.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:glad_tools/models/base64_image/base64_image.dart';
-import 'package:glad_tools/models/tool_object.dart';
-import 'package:glad_tools/models/json_beautify/json_beautify.dart';
-import 'package:glad_tools/models/jwt/jwt_parser.dart';
-import 'package:glad_tools/models/url_inspector/url_inspector.dart';
-import 'package:glad_tools/models/url_parser/url_parser.dart';
+import 'package:glad_tools/tools/base64_image/base64_image.dart';
+import 'package:glad_tools/tools/tool_object.dart';
+import 'package:glad_tools/tools/json_beautify/json_beautify.dart';
+import 'package:glad_tools/tools/jwt/jwt_parser.dart';
+import 'package:glad_tools/tools/url_inspector/url_inspector.dart';
+import 'package:glad_tools/tools/url_parser/url_parser.dart';
 
 const TOP_BAR_HEIGHT = 100.0;
 
@@ -39,14 +39,14 @@ class _MainViewState extends State<MainView> {
                       selected == e
                           ? const Icon(Icons.subdirectory_arrow_right)
                           : e.icon,
-                      Text(
+                      selected == e ? Text(
                         e.title,
                         style: TextStyle(
                           fontWeight: selected == e
                               ? FontWeight.bold
                               : FontWeight.normal,
                         ),
-                      ),
+                      ) : Container(),
                     ],
                   ),
                   onPressed: () => _selected(e),
